deploy/requirements.txt
-r ../requirements.txt
gunicorn
whitenoise
django-environ